requirements.txt
boto3
botocore
pytz
bs4
python-dotenv